Move JMXManagementModule to enterprise.
