+++ advanced/server-advanced/pom.xml
@@ -188,6 +188,11 @@
       <type>test-jar</type>
       <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.neo4j.app</groupId>
+      <artifactId>neo4j-server-enterprise</artifactId>
+      <version>3.0.0-SNAPSHOT</version>
+    </dependency>
 
   </dependencies>
