+++ enterprise/server-enterprise/src/main/java/org/neo4j/server/enterprise/jmx/ServerManagementMBean.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.server.advanced.jmx;
+package org.neo4j.server.enterprise.jmx;
 
 
 public interface ServerManagementMBean {
